requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

python-dotenv==1.0.1
pydantic==2.7.1
APScheduler==3.10.4

psycopg2
